hooks tests
