+++ test/index.js
@@ -7,7 +7,10 @@ var render = require("../create-element.js")
 var Node = require("../vtree/vnode")
 var TextNode = require("../vtree/vtext")
 var version = require("../vtree/version")
+var assertEqualDom = require("./lib/assert-equal-dom.js")
 
+require("./hook.js")
+require("./nested-properties.js")
 
 // VirtualNode tests
 test("Node is a function", function (assert) {
@@ -431,80 +434,6 @@ test("dom node remove", function (assert) {
     assert.end()
 })
 
-test("dom node style", function (assert) {
-
-
-    var a = h("div", {
-        style: {
-            border: "none",
-            className: "oops",
-            display: "none"
-        }
-    })
-
-    var b = h("div", {
-        style: {
-            border: "1px solid #000",
-            className: "oops",
-            display: ""
-        }
-    })
-
-    var rootNode = render(a)
-    assert.equal(rootNode.style.border, style("border", "none"))
-    assert.equal(rootNode.style.className, style("className", "oops"))
-    assert.equal(rootNode.style.display, style("display", "none"))
-    var s1 = rootNode.style
-    var equalNode = render(b)
-    assert.equal(equalNode.style.border, style("border", "1px solid #000"))
-    assert.equal(equalNode.style.className, style("className", "oops"))
-    assert.equal(equalNode.style.display, style("display", ""))
-    var newRoot = patch(rootNode, diff(a, b))
-    var s2 = newRoot.style
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.equal(newRoot.style.border, style("border", "1px solid #000"))
-    assert.equal(newRoot.style.className, style("className", "oops"))
-    assert.equal(newRoot.style.display, style("display", ""))
-    assert.equal(s1, s2)
-    assert.end()
-})
-
-test("dom node dataset", function (assert) {
-    var a = h("div", { dataset: { foo: "bar", bar: "oops" } })
-    var b = h("div", { dataset: { foo: "baz", bar: "oops" } })
-    var rootNode = render(a)
-    var d1 = rootNode.dataset
-    assert.equal(rootNode.dataset.foo, "bar")
-    assert.equal(rootNode.dataset.bar, "oops")
-    var equalNode = render(b)
-    var newRoot = patch(rootNode, diff(a, b))
-    var d2 = newRoot.dataset
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.equal(newRoot.dataset.foo, "baz")
-    assert.equal(newRoot.dataset.bar, "oops")
-    assert.equal(d1, d2)
-    assert.end()
-})
-
-test("dom node attributes", function (assert) {
-    if (!setAttributes()) {
-        assert.skip("No support for setting attributes array directly")
-        return assert.end()
-    }
-
-    var a = h("div", { attributes: { foo: "bar", bar: "oops" } })
-    var b = h("div", { attributes: { foo: "baz", bar: "oops" } })
-    var rootNode = render(a)
-    var a1 = rootNode.attributes
-    var equalNode = render(b)
-    var newRoot = patch(rootNode, diff(a, b))
-    var a2 = newRoot.attributes
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.equal(newRoot.attributes.foo, "baz")
-    assert.equal(newRoot.attributes.bar, "oops")
-    assert.equal(a1, a2)
-    assert.end()
-})
 
 test("reuse dom node without breaking", function (assert) {
     var hSpan = h("span", "hello")
@@ -873,110 +802,6 @@ test("Different namespaces creates a patch", function (assert) {
     assert.end()
 })
 
-test("Hooks are added to a hooks array on a node", function (assert) {
-    function Prop() {}
-    Prop.prototype.hook = function () {}
-
-    var node = new Node("div", {
-        "id": new Prop(),
-        "value": "not a hook"
-    }, [], null, null)
-
-    assert.equal(node.hooks.length, 1)
-    assert.equal(node.hooks[0], "id")
-    assert.equal(node.descendantHooks, false)
-    assert.end()
-})
-
-test("Node child hooks are identified", function (assert) {
-    function Prop() {}
-    Prop.prototype.hook = function () {}
-
-    var node = new Node("div", {
-        "id": new Prop(),
-        "value": "not a hook"
-    }, [], null, null)
-
-    var parentNode = new Node("div", {
-        "id": "not a hook"
-    }, [node], null, null)
-
-    assert.equal(node.hooks.length, 1)
-    assert.equal(node.hooks[0], "id")
-    assert.equal(parentNode.hooks, null)
-    assert.ok(parentNode.descendantHooks)
-    assert.end()
-})
-
-// Safely translates style values using the DOM in the browser
-function style(name, value) {
-    var node = render(h())
-    node.style[name] = value
-    return node.style[name]
-}
-
-function assertEqualDom(assert, a, b) {
-    assert.ok(areEqual(a, b) && areEqual(b, a), "Dom structures are equal")
-}
-
-function areEqual(a, b) {
-    for (var key in a) {
-        if (key !== "parentNode" &&
-            key !== "parentElement" &&
-            key !== "defaultView" &&
-            key !== "ownerElement" &&
-            key !== "nextElementSibling" &&
-            key !== "nextSibling" &&
-            key !== "previousElementSibling" &&
-            key !== "previousSibling" &&
-            key !== "document" &&
-            key !== "window" &&
-            key !== "frames" &&
-            key !== "top" &&
-            key !== "parent" &&
-            key !== "self" &&
-            key !== "outerHTML" &&
-            key !== "innerHTML"
-        ) {
-            if (key === "ownerDocument") return a[key] === b[key]
-            if (key === "style") return equalStyle(a[key], b[key])
-            if (typeof a === "object" || typeof a === "function") {
-                if (!areEqual(a[key], b[key])) {
-                    return false
-                }
-            } else {
-                if (a !== b) {
-                    return false
-                }
-            }
-        }
-    }
-
-    return true
-}
-
-// CssStyleDeclaration indexes the styles, which could be out of order
-// This is a left sided check. Note that we call equal(a, b) and equal(b, a)
-function equalStyle(a, b) {
-    var keys = []
-    for (var key in a) {
-        if ("" + parseInt(key, 10) === key) {
-            continue
-        } else {
-            keys.push(key)
-        }
-    }
-
-    keys.sort()
-
-    for (var i = 0; i < keys.length; i++) {
-        if (a[key] !== b[key]) {
-            return false
-        }
-    }
-
-    return true
-}
 
 function patchCount(patch) {
     var count = 0
@@ -990,17 +815,9 @@ function patchCount(patch) {
     return count
 }
 
-function setAttributes() {
+// Safely translates style values using the DOM in the browser
+function style(name, value) {
     var node = render(h())
-    var attr = "foo"
-    try {
-        if (!("attributes" in node)) {
-            node.attributes = {}
-        }
-
-        node.attributes[attr] = "bar"
-        return true
-    } catch (e) {
-        return false
-    }
+    node.style[name] = value
+    return node.style[name]
 }
